requirements.txt
asgiref==3.2.10
click==7.1.2
cmake==3.17.3
Django==3.0.8
django-clear-cache==0.3
django-webcam==0.1.4b2
Flask==1.1.2
flaskwebgui==0.1.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.0
opencv-python==4.3.0.36
Pillow==7.2.0
psutil==5.7.0
pytz==2020.1
sqlparse==0.3.1
Werkzeug==1.0.1
whichcraft==0.6.1
